==> r/6_barplot_tables.Rmd <==
---
title: "barplot_tables"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

To break down the family membership of ASVs by host and habitat.

## Libraries

```{r}

library(reshape2)

```

## Things Needed

```{r}

# data frames of unique ASVs
readRDS("../intermediate_files/rds/ak_total_unique_rbind.rds")
readRDS("../intermediate_files/rds/ro_total_unique_rbind.rds")

# data frames of shared ASVs
readRDS("../intermediate_files/rds/ak_total_shared_rbind.rds")
readRDS("../intermediate_files/rds/ro_total_shared_rbind.rds")

# matrix of ASV counts by host
readRDS("../intermediate_files/rds/ak_full_mat.rds")

# matrix of ASV counts by host
readRDS("../intermediate_files/rds/ro_full_mat.rds")

```

# Unique

## AK unique ASVs

```{r}

# make new copy of unique ak df
ak_unique_table_df <- ak_total_unique_rbind

# group by Sample (host/soil) and family
ak_unique_table_df <- ak_unique_table_df %>% group_by(Host, Family) %>%
  # calculate % of unique ASVs per host/soil belonging to a certain family
  summarise(count=round((length(ASV)/nrow(ak_full_mat) * 100), digits=2)) %>%
  # use dcast to "unmelt" df and have families as row names and Samples as col names with counts as the cell values
  dcast(Family ~ Host) %>%
  column_to_rownames('Family') %>%
  # rename columns
  rename_all(function(x) paste("AK", x, sep=" "))

# Changes NAs to 0
ak_unique_table_df[is.na(ak_unique_table_df)] <- 0


```


## RO Unique ASVs

```{r}

# make new copy of melted unique ak df
ro_unique_table_df <- ro_total_unique_rbind

# group by Sample (host/soil) and family
ro_unique_table_df <- ro_unique_table_df %>% group_by(Host, Family) %>%
  # calculate % of unique ASVs per host/soil belonging to a certain family
  summarise(count=round((length(ASV)/nrow(ro_full_mat) * 100), digits=2)) %>%
  # use dcast to "unmelt" df and have families as row names and Samples as col names with counts as the cell values
  dcast(Family ~ Host) %>%
  column_to_rownames('Family') %>%
  # rename columns
  rename_all(function(x) paste("RO", x, sep=" "))

# Changes NAs to 0
ro_unique_table_df[is.na(ro_unique_table_df)] <- 0

```


## Merge

```{r}

# merge
merged_unique_table <- merge(ak_unique_table_df, ro_unique_table_df, by="row.names")

# make families into row names
merged_unique_table <- merged_unique_table %>% column_to_rownames('Row.names')

# reorder cols. currently they're grouped by habitat, this groups them by sample
merged_unique_table <- merged_unique_table[c(1,8,2,9,3,10,4,11,5,12,6,13,7,14)]


```


# Shared

## AK shared

```{r}

# make new copy of melted unique ak df
ak_shared_table_df <- ak_total_shared_rbind

# group by family and summarise to show number of unique ASVs per family
# every sample has the same number of shared ASVs
ak_shared_table_df <- ak_shared_table_df %>% group_by(Family) %>%
  # calculates the % of shared ASVs belonging to a certain family
  summarise(count=round((length(unique(ASV))/nrow(ak_full_mat) * 100), digits=2)) %>%
  column_to_rownames('Family') %>%
  # rename column
  rename_all(function(x) paste("AK"))


```


## RO Shared

```{r}

# make new copy of melted unique ak df
ro_shared_table_df <- ro_total_shared_rbind

# group by family and summarise to show number of unique ASVs per family
# every sample has the same number of shared ASVs
ro_shared_table_df <- ro_shared_table_df %>% group_by(Family) %>%
  # calculates the % of shared ASVs belonging to a certain family
  summarise(count=round((length(unique(ASV))/nrow(ro_full_mat) * 100), digits=2)) %>%
  column_to_rownames('Family') %>%
  # rename column
  rename_all(function(x) paste("RO"))

```


## Shared Merge

```{r}

#merge
merged_shared_table <- merge(ak_shared_table_df, ro_shared_table_df, by="row.names")

# make families into row names
merged_shared_table <- merged_shared_table %>% column_to_rownames('Row.names')

```

## Export

```{r}

write.csv(merged_unique_table, file="../outputs/unique_table.csv")

write.csv(merged_shared_table, file="../outputs/shared_table.csv")

```
