==> r/8_permanova.Rmd <==
---
title: "permanova"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Data prep

```{r}

library(vegan)

# the otu table
nmds_abun 

# by habitat
adon_habitat <- adonis(nmds_abun ~ data_scores$Habitat, permutations = 9999, method = "bray")

# by host
adon_host <- adonis(nmds_abun ~ data_scores$Sample, permutations = 9999, method = "bray")

# by habitat + host
adon_hab_and_host <- adonis(nmds_abun ~ data_scores$Habitat + data_scores$Sample, permutations = 9999, method = "bray")


```


## Export

```{r}

capture.output(adon_habitat, file="outputs/permanova_habitat.txt")

capture.output(adon_host, file="outputs/permanova_host.txt")

capture.output(adon_hab_and_host, file="outputs/permanova_habitat_and_host.txt")


```
